Break out the `schema ...` hierarchy
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`schema ...` hierarchy.
